+++ tests/integration/chat/v2/service/channel/test_message.py
@@ -39,11 +39,50 @@ class MessageTestCase(IntegrationTestCase):
                 "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "date_created": "2016-03-24T20:37:57Z",
                 "date_updated": "2016-03-24T20:37:57Z",
+                "last_updated_by": null,
                 "was_edited": false,
                 "from": "system",
                 "attributes": "{}",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
+                "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+            }
+            '''
+        ))
+
+        actual = self.client.chat.v2.services(sid="ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
+                                    .channels(sid="CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa") \
+                                    .messages(sid="IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa").fetch()
+
+        self.assertIsNotNone(actual)
+
+    def test_fetch_media_response(self):
+        self.holodeck.mock(Response(
+            200,
+            '''
+            {
+                "sid": "IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "service_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "to": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                "date_created": "2016-03-24T20:37:57Z",
+                "date_updated": "2016-03-24T20:37:57Z",
+                "last_updated_by": null,
+                "was_edited": false,
+                "from": "system",
+                "attributes": "{}",
+                "body": "Hello",
+                "index": 0,
+                "type": "media",
+                "media": {
+                    "sid": "MEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                    "size": 99999999999999,
+                    "content_type": "application/pdf",
+                    "filename": "hello.pdf"
+                },
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
@@ -86,10 +125,13 @@ class MessageTestCase(IntegrationTestCase):
                 "attributes": null,
                 "date_created": "2016-03-24T20:37:57Z",
                 "date_updated": "2016-03-24T20:37:57Z",
+                "last_updated_by": "system",
                 "was_edited": false,
                 "from": "system",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
@@ -101,7 +143,7 @@ class MessageTestCase(IntegrationTestCase):
 
         self.assertIsNotNone(actual)
 
-    def test_create_with_attributes_response(self):
+    def test_create_with_all_response(self):
         self.holodeck.mock(Response(
             201,
             '''
@@ -111,13 +153,16 @@ class MessageTestCase(IntegrationTestCase):
                 "service_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "to": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "date_created": "2016-03-24T20:37:57Z",
-                "date_updated": "2016-03-24T20:37:57Z",
-                "was_edited": false,
+                "date_created": "2015-12-16T22:18:37Z",
+                "date_updated": "2015-12-16T22:18:38Z",
+                "last_updated_by": "username",
+                "was_edited": true,
                 "from": "system",
-                "attributes": "{}",
+                "attributes": "{\\"test\\": \\"test\\"}",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''
@@ -165,11 +210,37 @@ class MessageTestCase(IntegrationTestCase):
                         "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                         "date_created": "2016-03-24T20:37:57Z",
                         "date_updated": "2016-03-24T20:37:57Z",
+                        "last_updated_by": null,
                         "was_edited": false,
                         "from": "system",
                         "attributes": "{}",
                         "body": "Hello",
                         "index": 0,
+                        "type": "text",
+                        "media": null,
+                        "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
+                    },
+                    {
+                        "sid": "IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "account_sid": "ACaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "service_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "to": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                        "date_created": "2016-03-24T20:37:57Z",
+                        "date_updated": "2016-03-24T20:37:57Z",
+                        "last_updated_by": null,
+                        "was_edited": false,
+                        "from": "system",
+                        "attributes": "{}",
+                        "body": "Hello",
+                        "index": 0,
+                        "type": "media",
+                        "media": {
+                            "sid": "MEaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
+                            "size": 99999999999999,
+                            "content_type": "application/pdf",
+                            "filename": "hello.pdf"
+                        },
                         "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
                     }
                 ]
@@ -256,13 +327,16 @@ class MessageTestCase(IntegrationTestCase):
                 "service_sid": "ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "to": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
                 "channel_sid": "CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa",
-                "attributes": "{\\"test\\": \\"test\\"}",
-                "date_created": "2016-03-24T20:37:57Z",
-                "date_updated": "2016-03-24T20:37:57Z",
-                "was_edited": false,
+                "attributes": "{ \\"foo\\": \\"bar\\" }",
+                "date_created": "2015-12-16T22:18:37Z",
+                "date_updated": "2015-12-16T22:18:38Z",
+                "last_updated_by": "username",
+                "was_edited": true,
                 "from": "system",
                 "body": "Hello",
                 "index": 0,
+                "type": "text",
+                "media": null,
                 "url": "https://chat.twilio.com/v2/Services/ISaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Channels/CHaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa/Messages/IMaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa"
             }
             '''