+++ twilio/__init__.py
@@ -1,3 +1,3 @@
 
-__version_info__ = ('6', '5', '0')
+__version_info__ = ('6', '6', '1')
 __version__ = '.'.join(__version_info__)