+++ twilio/http/http_client.py
@@ -1,6 +1,6 @@
 from requests import Request, Session
 
-from twilio.http import HttpClient, get_cert_file
+from twilio.http import HttpClient
 from twilio.http.response import Response
 from twilio.http.request import Request as TwilioRequest
 
@@ -9,15 +9,10 @@ class TwilioHttpClient(HttpClient):
     """
     General purpose HTTP Client for interacting with the Twilio API
     """
-    def __init__(self, connection_pool=True):
+    def __init__(self, pool_connections=True):
+        self.session = Session() if pool_connections else None
         self.last_request = None
         self.last_response = None
-
-        if connection_pool:
-            self.session = Session()
-            self.session.verify = get_cert_file()
-        else:
-            self.session = None
 
     def request(self, method, url, params=None, data=None, headers=None, auth=None, timeout=None,
                 allow_redirects=False):
@@ -39,11 +34,7 @@ class TwilioHttpClient(HttpClient):
         """
 
         self.last_response = None
-        session = self.session
-        if session is None:
-            session = Session()
-            session.verify = get_cert_file()
-
+        session = self.session or Session()
         request = Request(method.upper(), url, params=params, data=data, headers=headers, auth=auth)
         self.last_request = TwilioRequest(method.upper(), url, auth=auth, params=params, data=data,
                                           headers=headers)