+++ twilio/rest/video/v1/room/__init__.py
@@ -37,7 +37,8 @@ class RoomList(ListResource):
     def create(self, enable_turn=values.unset, type=values.unset,
                unique_name=values.unset, status_callback=values.unset,
                status_callback_method=values.unset, max_participants=values.unset,
-               record_participants_on_connect=values.unset):
+               record_participants_on_connect=values.unset,
+               video_codecs=values.unset, media_region=values.unset):
         """
         Create a new RoomInstance
 
@@ -48,6 +49,8 @@ class RoomList(ListResource):
         :param unicode status_callback_method: The status_callback_method
         :param unicode max_participants: The max_participants
         :param bool record_participants_on_connect: The record_participants_on_connect
+        :param RoomInstance.VideoCodec video_codecs: The video_codecs
+        :param unicode media_region: The media_region
 
         :returns: Newly created RoomInstance
         :rtype: twilio.rest.video.v1.room.RoomInstance
@@ -60,6 +63,8 @@ class RoomList(ListResource):
             'StatusCallbackMethod': status_callback_method,
             'MaxParticipants': max_participants,
             'RecordParticipantsOnConnect': record_participants_on_connect,
+            'VideoCodecs': video_codecs,
+            'MediaRegion': media_region,
         })
 
         payload = self._version.create(
@@ -380,6 +385,10 @@ class RoomInstance(InstanceResource):
     class RoomType(object):
         PEER_TO_PEER = "peer-to-peer"
         GROUP = "group"
+
+    class VideoCodec(object):
+        VP8 = "VP8"
+        H264 = "H264"
 
     def __init__(self, version, payload, sid=None):
         """